Extract lib/zendesk_api/resource_user.rb
